Just a bunch more changes to an engine that still isn't very functional

+++ RheenCore/RheenCore/src/rheencore/input/Mouse.java
@@ -6,16 +6,16 @@ import java.awt.event.MouseMotionListener;
 import java.awt.event.MouseWheelEvent;
 import java.awt.event.MouseWheelListener;
 
-import rheencore.maths.vector.f.Vec2f;
+import rheencore.maths.vector.d.Vec2d;
 
 public class Mouse implements MouseListener, MouseMotionListener, MouseWheelListener{
 
-	public static Vec2f position;
-	public static Vec2f moved;
+	public static Vec2d position;
+	public static Vec2d moved;
 	
 	public Mouse(){
-		position = new Vec2f();
-		moved = new Vec2f();
+		position = new Vec2d();
+		moved = new Vec2d();
 	}
 	
 	@Override
@@ -27,7 +27,7 @@ public class Mouse implements MouseListener, MouseMotionListener, MouseWheelList
 	@Override
 	public void mouseMoved(MouseEvent e)
 	{
-		Mouse.position.set(new Vec2f(e.getX(), e.getY()));
+		Mouse.position.set(new Vec2d(e.getX(), e.getY()));
 		
 		e.consume();
 	}
